revert: revert remove wallet connect page (#8466)
