+++ src/assets/translations/en/main.json
@@ -301,21 +301,8 @@
     "notSupported": "%{asset} is not supported by %{wallet}"
   },
   "connectWalletPage": {
-    "title": "The Original Multichain Exchange",
-    "primaryTitle": "Welcome to Multichain DeFi.",
-    "secondaryTitle": "Private. Decentralized. Non-custodial.",
-    "primaryDescription": "Trade, bridge, and earn rewards effortlessly. Experience Native Bitcoin, Dogecoin, and more. Manage Liquidity and DeFi Positions in One-Click.",
-    "snapDescription": "Download the ShapeShift Multichain Snap and Unlock Bitcoin, Dogecoin, and more for your MetaMask.",
     "shapeshift": "ShapeShift",
-    "exploreThe": "Explore the",
-    "defiUniverse": "DeFi Universe",
-    "body": "Trade, bridge & earn. Private. Community owned. Non-custodial. Decentralized.",
-    "header": "Please connect a wallet to get started",
-    "cta": "Connect or Create Wallet",
-    "viewADemo": "View a Demo",
-    "dontHaveWallet": "Don't have a wallet?",
     "welcomeBack": "Welcome back!",
-    "welcomeBody": "Connecting your wallet is like “logging in”. Select your wallet from the options to get started.",
     "welcomeToShapeShift": "Welcome to ShapeShift",
     "mobileWelcomeBody": "Create a brand new wallet or add an existing one to get started easily.",
     "mobileSelectBody": "Select from any of your saved wallets below, or create or import a new wallet.",
