+++ src/components/Modals/Settings/ClearCache.tsx
@@ -72,7 +72,9 @@ export const ClearCache = ({ appHistory }: ClearCacheProps) => {
       await persistor.purge()
       // send them back to the connect wallet route in case the bug was something to do with the current page
       // and so they can reconnect their native wallet to avoid the app looking broken in an infinite loading state
-      appHistory.replace('/connect-wallet')
+      if (isMobileApp) {
+        appHistory.replace('/connect-mobile-wallet')
+      }
       // reload the page
       isMobileApp ? reloadWebview() : window.location.reload()
     } catch (e) {}
